Add meta tags and adjust UI elements
Two new meta-tags have been added to the HTML head for "naive-ui-style" and "vueuc-style". In addition, some user interface elements have been adjusted in the HomeView.vue and the AIChatWindow.vue files to make the chat window sticky at the top and to improve the theme and placement of the input fields. Also, the BackgroundCards.vue file was updated to use the darkTheme from Naive-ui.